added-the-extract_info-function
This part of the code takes care of creating a new extract_info function that takes care of extracting info from the file that we are analyzing. Remains to be done the creation of the new parts of the function that take care of extracting information related to version, load configurations, etc. Also added the requirements .txt file that contains the libraries to be installed in order to run the application from a sandbox environment.