add-entropy-function
This part of the code takes care of generating the entropy of data based on Shannon entropy formula. After the implementation of this function, the function finds usage in the get_resources function in terms of finding the entropy of data and adding key-value pairs of structure_size -> entropy to the resources dictionary.